Removed dead code.
