+++ app/services/lab/lims/order_serializer.rb
@@ -20,6 +20,7 @@ module Lab
             order_location: format_order_location(serialized_order.encounter_id),
             sample_type: format_sample_type(serialized_order.specimen.name),
             sample_status: format_sample_status(serialized_order.specimen.name),
+            sample_statuses: format_sample_status_trail(order),
             districy: current_district, # yes districy [sic]...
             priority: serialized_order.reason_for_test.name,
             date_created: serialized_order.order_date,
@@ -68,6 +69,28 @@ module Lab
 
         def format_sample_status(name)
           name.casecmp?('Unknown') ? 'specimen_not_collected' : 'specimen_collected'
+        end
+
+        def format_sample_status_trail(order)
+          if order.concept_id == ConceptName.find_by_name!('Unknown').concept_id
+            return []
+          end
+
+          user = User.find(order.discontinued_by || order.creator)
+          drawn_by = PersonName.find_by(user.user_id)
+          drawn_date = order.discontinued_date || order.start_date
+
+          [
+            drawn_date.strftime('%Y%m%d%H%M%S') => {
+              'status' => 'Drawn',
+              'updated_by' => {
+                'first_name' => drawn_by&.given_name || user.username,
+                'last_name' => drawn_by&.family_name,
+                'phone_number' => nil,
+                'id' => user.username
+              }
+            }
+          ]
         end
 
         def format_test_results(order)
